refactor(#26): InventoryResponse dto/response로 이동
