+++ src/main/java/software_capstone/backend/app/inventory/controller/InventoryController.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import software_capstone.backend.app.inventory.dto.InventoryResponse;
+import software_capstone.backend.app.inventory.dto.response.InventoryResponse;
 import software_capstone.backend.app.inventory.service.InventoryService;
 
 import java.util.List;
