+++ src/main/java/software_capstone/backend/app/inventory/controller/InventoryControllerDocs.java
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.http.ResponseEntity;
-import software_capstone.backend.app.inventory.dto.InventoryResponse;
+import software_capstone.backend.app.inventory.dto.response.InventoryResponse;
 
 import java.util.List;
